broken tests into multiple files + change board test list behavior (each run create a new list)
