+++ test/test_board.py
@@ -1,104 +1,9 @@
+#!/usr/bin/python
+from __future__ import with_statement, print_function
 from datetime import datetime, timedelta
-from trello import TrelloClient, Unauthorized, ResourceUnavailable
 import unittest
 import os
-
-
-class TrelloClientTestCase(unittest.TestCase):
-    """
-
-    Tests for TrelloClient API. Note these test are in order to
-    preserve dependencies, as an API integration cannot be tested
-    independently.
-
-    """
-
-    def setUp(self):
-        self._trello = TrelloClient(os.environ['TRELLO_API_KEY'],
-                                    token=os.environ['TRELLO_TOKEN'])
-
-    def test01_list_boards(self):
-        self.assertEquals(
-            len(self._trello.list_boards()),
-            int(os.environ['TRELLO_TEST_BOARD_COUNT']))
-
-    def test10_board_attrs(self):
-        boards = self._trello.list_boards()
-        for b in boards:
-            self.assertIsNotNone(b.id, msg="id not provided")
-            self.assertIsNotNone(b.name, msg="name not provided")
-            self.assertIsNotNone(b.description, msg="description not provided")
-            self.assertIsNotNone(b.closed, msg="closed not provided")
-            self.assertIsNotNone(b.url, msg="url not provided")
-
-    def test20_board_all_lists(self):
-        boards = self._trello.list_boards()
-        for b in boards:
-            try:
-                b.all_lists()
-            except Exception:
-                self.fail("Caught Exception getting lists")
-
-    def test21_board_open_lists(self):
-        boards = self._trello.list_boards()
-        for b in boards:
-            try:
-                b.open_lists()
-            except Exception:
-                self.fail("Caught Exception getting open lists")
-
-    def test22_board_closed_lists(self):
-        boards = self._trello.list_boards()
-        for b in boards:
-            try:
-                b.closed_lists()
-            except Exception:
-                self.fail("Caught Exception getting closed lists")
-
-    def test30_list_attrs(self):
-        boards = self._trello.list_boards()
-        for b in boards:
-            for l in b.all_lists():
-                self.assertIsNotNone(l.id, msg="id not provided")
-                self.assertIsNotNone(l.name, msg="name not provided")
-                self.assertIsNotNone(l.closed, msg="closed not provided")
-            break  # only need to test one board's lists
-
-    def test50_list_cards(self):
-        boards = self._trello.list_boards()
-        for b in boards:
-            for l in b.all_lists():
-                for c in l.list_cards():
-                    self.assertIsNotNone(c.id, msg="id not provided")
-                    self.assertIsNotNone(c.name, msg="name not provided")
-                    self.assertIsNotNone(c.description,
-                                         msg="description not provided")
-                    self.assertIsNotNone(c.closed, msg="closed not provided")
-                    self.assertIsNotNone(c.url, msg="url not provided")
-                break
-            break
-        pass
-
-    def test51_fetch_cards(self):
-        """
-        Tests fetching all attributes for all cards
-        """
-        boards = self._trello.list_boards()
-        for b in boards:
-            for l in b.all_lists():
-                for c in l.list_cards():
-                    c.fetch()
-
-                    self.assertIsInstance(c.date_last_activity, datetime,
-                                          msg='date not provided')
-                    self.assertTrue(len(c.board_id) > 0,
-                                    msg='board id not provided')
-                break
-            break
-        pass
-
-    def test52_list_hooks(self):
-        self.assertIsInstance(self._trello.list_hooks(), list)
+from trello import TrelloClient, Unauthorized, ResourceUnavailable
 
 
 class TrelloBoardTestCase(unittest.TestCase):
@@ -116,19 +21,20 @@ class TrelloBoardTestCase(unittest.TestCase):
             if b.name == os.environ['TRELLO_TEST_BOARD_NAME']:
                 cls._board = b
                 break
-        try:
-            cls._list = cls._board.open_lists()[0]
-        except IndexError:
-            cls._list = cls._board.add_list('List')
+        if not cls._board:
+            cls.fail("Couldn't find test board")
+        cls._list = cls._board.add_list(str(datetime.now()))
 
     def _add_card(self, name, description=None):
-        card = self._list.add_card(name, description)
-        self.assertIsNotNone(card, msg="card is None")
-        self.assertIsNotNone(card.id, msg="id not provided")
-        self.assertEquals(card.name, name)
-        self.assertEqual(card.board, self._board)
-        self.assertEqual(card.trello_list, self._list)
-        return card
+        try:
+            card = self._list.add_card(name, description)
+            self.assertIsNotNone(card, msg="card is None")
+            self.assertIsNotNone(card.id, msg="id not provided")
+            self.assertEquals(card.name, name)
+            return card
+        except Exception as e:
+            print(str(e))
+            self.fail("Caught Exception adding card")
 
     def test40_add_card(self):
         name = "Testing from Python - no desc"
@@ -206,8 +112,6 @@ class TrelloBoardTestCase(unittest.TestCase):
             else:
                 self.fail(msg='Unexpected card found')
 
-            self.assertFalse(hasattr(card, 'trello_list'))
-
         self.assertIsInstance(self._board.all_cards(), list)
         self.assertIsInstance(self._board.open_cards(), list)
         self.assertIsInstance(self._board.closed_cards(), list)
@@ -227,8 +131,6 @@ class TrelloBoardTestCase(unittest.TestCase):
         card.fetch()
         actual_due_date = card.due[:10]
         self.assertEquals(expected_due_date, actual_due_date)
-        # Note that set_due passes only the date, stripping time
-        self.assertEquals(card.due_date.date(), due_date.date())
 
     def test53_checklist(self):
         name = "Testing from Python"
@@ -252,21 +154,6 @@ class TrelloBoardTestCase(unittest.TestCase):
         card.set_description(description)
         self.assertEquals(card.name, name)
         self.assertEquals(card.description, description)
-
-    def test55_set_pos(self):
-        card_names = lambda: [c.name for c in self._list.list_cards()]
-        self._list.add_card('card1')
-        card2 = self._list.add_card('card2')
-        names = card_names()
-        self.assertGreater(names.index('card2'), names.index('card1'))
-
-        card2.set_pos('top')
-        names = card_names()
-        self.assertGreater(names.index('card1'), names.index('card2'))
-
-        card2.set_pos('bottom')
-        names = card_names()
-        self.assertGreater(names.index('card2'), names.index('card1'))
 
     def test60_delete_cards(self):
         cards = self._board.get_cards()
@@ -311,9 +198,9 @@ class TrelloBoardTestCase(unittest.TestCase):
 
 
 def suite():
-    tests = ['test01_list_boards', 'test10_board_attrs', 'test20_add_card']
-    return unittest.TestSuite(map(TrelloClientTestCase, tests))
-
+    # tests = ['test01_list_boards', 'test10_board_attrs', 'test20_add_card']
+    # return unittest.TestSuite(map(TrelloBoardTestCase, tests))
+    return unittest.TestLoader().loadTestsFromTestCase(TrelloBoardTestCase)
 
 if __name__ == "__main__":
     unittest.main()
